added delete route to active donations

+++ src/page-components/components/RemoveBtn.jsx
@@ -6,13 +6,13 @@ import { useRouter } from "next/navigation";
 export default function RemoveBtn({ id }) {
   const router = useRouter();
   const removeItem = async () => {
-    const confirmed = confirm("Are you sure?");
+   const confirmed = confirm("Are you sure?");
 
     if (confirmed) {
-      const res = await fetch(`http://localhost:3000/api/items?id=${id}`, {
+      const res = await fetch(`http://localhost:3000/api/items/${id}/page`, {
         method: "DELETE",
       });
-
+      
       if (res.ok) {
         router.refresh();
       }
